Move testing data to fakes file
